requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.1
crispy-bootstrap4==2023.1
Django==4.2.6
django-crispy-forms==2.0
django-rest-passwordreset==1.3.0
idna==3.4
Pillow==10.1.0
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.0.7